AIS-API/pyproject.toml
[project]
name = "ais-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]==0.116.1",
    "numpy==1.24",
    "onnxruntime>=1.22.1",
    "pillow==11.2.1",
    "rembg==2.0.61",
    "requests>=2.32.4",
    "tensorflow==2.10.0",
    "tqdm>=4.67.1",
]
